Fixed a bug in inOrderTraversal
We were attempting to print a leaf node in the recursive forloop even when the node in question was a leaf node, resulting in an exception. In order to avoid this, we check to see if it is a leaf node and do the forloop without the recursive calls to the method.