Added Constructors and an Inner Class
We need to have 4 constructors; the creation constructor needs to have the option for with or without the cache, and so does the read constructor (which constructs the BTree from a BTree file). Because these are two separate constructors, we will be adding two more constructors, so we have a total of 4.

Also, I realized that the Cache needs to be an inner class within the BTree class, but I can't seem to get it working correctly. The Cache class doesn't work right now, but what I wrote is in there.